ai-service/requirements.txt
tensorflow>=2.10.0,<2.16.0
tf_keras>=2.10.0
Pillow>=10.0.0
numpy>=1.23.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
matplotlib>=3.7.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
